+++ src/router/lsof/tests/LTszoff.c
@@ -357,7 +357,7 @@ testlsof(tt, opt, xval)
  */
     if ((cem = ConvStatDev(&sb.st_dev, &stdc)))
 	PrtMsgX(buf, Pn, cleanup, 1);
-    (void) snprintf(ibuf, sizeof(ibuf) - 1, "%u", (unsigned int)sb.st_ino);
+    (void) snprintf(ibuf, sizeof(ibuf) - 1, "%" PRIu64, (uint64_t)sb.st_ino);
     ibuf[sizeof(ibuf) - 1] = '\0';
 /*
  * Complete the option vector and start lsof execution.
